Handle header attribute on __init_subclass__
Now the __init_subclass__ method from genericstruct handle the creation of
subclasses, specially by defining the need of a 'message_type' attribute on the
class definition and also handling the creation of the header attribute of the
messages, assuring the version of the header and the message_type to be defined
on the header. this also allowed us to remove this part of the code from the
metaclass __new__ method, making it simpler.