GenericMessage now knows all messages classes.
Now the GenericMessage class will hold the _messages attributes, that will keep
a record of all messages created in a form of a dict, grouped by 'of_version'.

There is also a change on the __init_subclass__ method from GenericMessage to
avoid users from accessing the `_messages` attribute from the subclasses.